fixed Long transaction id to int

+++ back_end/src/main/java/back/data/BookingDAOImpl.java
@@ -21,8 +21,8 @@ public class BookingDAOImpl implements BookingDAO {
     }
 
     @Override
-    public long bookRoomForVisitor(User user, Room room, String sqlStartDate, String sqlEndDate, int occupants) throws JTHDataBaseException {
-        long transac_id = dataAccess.insertTransaction(user, room, sqlStartDate, sqlEndDate, occupants);
+    public int bookRoomForVisitor(User user, Room room, String sqlStartDate, String sqlEndDate, int occupants) throws JTHDataBaseException {
+        int transac_id = dataAccess.insertTransaction(user, room, sqlStartDate, sqlEndDate, occupants);
         if (transac_id>0) search.pushTransaction(room.getId(), new Transaction(user.getId(), room.getId(), -1, sqlStartDate, sqlEndDate, -1.0, occupants));
         return transac_id;
     }
