+++ back_end/src/main/java/back/data/jdbc/DataAccess.java
@@ -333,7 +333,7 @@ public class DataAccess {
         }
     }
 
-    public long insertTransaction(User user, Room room, String sqlStartDate, String sqlEndDate, int occupants) throws JTHDataBaseException {
+    public int insertTransaction(User user, Room room, String sqlStartDate, String sqlEndDate, int occupants) throws JTHDataBaseException {
         try {
             Boolean res = transactionTemplate.execute(status -> {
                 if (occupants > room.getMaxOccupants()){
@@ -361,7 +361,7 @@ public class DataAccess {
             e.printStackTrace();
             throw new JTHDataBaseException();
         }
-        return jdbcTemplate.queryForObject("select id from transactions where room_id = ? and start_date = ? ::date and end_date = ? ::date", new Object[]{room.getId(), sqlStartDate, sqlEndDate}, Long.class);
+        return jdbcTemplate.queryForObject("select id from transactions where room_id = ? and start_date = ? ::date and end_date = ? ::date", new Object[]{room.getId(), sqlStartDate, sqlEndDate}, Integer.class);
     }
 
     public int countTransactions(Room room, String sqlStartDate, String sqlEndDate) throws JTHDataBaseException {
